OSIS-5372 moved ProgramTreeIdentityFactory into a separate file
